add logging to delete forms task (#1767)
* add logging to delete forms task

* add specific ids to be deleted

* add log and return for empty query

+++ nest-forms-backend/src/forms/subservices/forms-task.subservice.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common'
+import { Injectable, Logger } from '@nestjs/common'
 import { Cron, CronExpression } from '@nestjs/schedule'
 import { FormState, Prisma } from '@prisma/client'
 
@@ -7,6 +7,8 @@ import PrismaService from '../../prisma/prisma.service'
 
 @Injectable()
 export default class FormsTaskSubservice {
+  private logger: Logger = new Logger(FormsTaskSubservice.name)
+
   constructor(
     private readonly prismaService: PrismaService,
     private readonly filesService: FilesService,
@@ -14,6 +16,7 @@ export default class FormsTaskSubservice {
 
   @Cron(CronExpression.EVERY_DAY_AT_1AM)
   async deleteOldDraftForms(): Promise<void> {
+    this.logger.log('Deleting old draft forms.')
     const oneWeekAgo = new Date(Date.now() - 1000 * 60 * 60 * 24 * 7)
 
     // Get the forms that are in DRAFT state, older than one week and have no data.
@@ -36,17 +39,29 @@ export default class FormsTaskSubservice {
         },
       },
     })
+    if (forms.length === 0) {
+      this.logger.log('No old draft forms found to delete.')
+      return
+    }
+
     const formIds = forms.map((form) => form.id)
     const fileIds = forms.flatMap((form) => form.files.map((file) => file.id))
+    this.logger.log(
+      `Found ${formIds.length} old draft forms: ${formIds.join(', ')}`,
+    )
+    this.logger.log(
+      `Found ${fileIds.length} files to delete from old draft forms: ${fileIds.join(', ')}`,
+    )
 
     // Delete the forms along with its files
     await this.filesService.deleteFileMany(fileIds)
-    await this.prismaService.forms.deleteMany({
+    const deleted = await this.prismaService.forms.deleteMany({
       where: {
         id: {
           in: formIds,
         },
       },
     })
+    this.logger.log(`Deleted ${deleted.count} old draft forms.`)
   }
 }
